remove coin handler
