+++ vm/database/db.go
@@ -7,7 +7,6 @@ type Visitor struct {
 	MapHandler
 	ContractHandler
 	TokenHandler
-	CoinHandler
 	RollbackHandler
 	DelaytxHandler
 }
@@ -21,7 +20,6 @@ func NewVisitor(cacheLength int, cb IMultiValue) *Visitor {
 		BasicHandler:    BasicHandler{cachedDB},
 		MapHandler:      MapHandler{cachedDB},
 		ContractHandler: ContractHandler{cachedDB},
-		CoinHandler:     CoinHandler{cachedDB},
 		TokenHandler:    TokenHandler{cachedDB},
 		GasHandler:      GasHandler{cachedDB},
 	}
@@ -50,7 +48,6 @@ func NewBatchVisitor(lruDB *LRU) (*Visitor, Mapper) {
 		BasicHandler:    BasicHandler{watcher},
 		MapHandler:      MapHandler{watcher},
 		ContractHandler: ContractHandler{watcher},
-		CoinHandler:     CoinHandler{watcher},
 		TokenHandler:    TokenHandler{watcher},
 		GasHandler:      GasHandler{watcher},
 	}
