+++ vm/engine.go
@@ -26,7 +26,7 @@ const (
 
 var (
 	errSetUpArgs = errors.New("key does not exist")
-	errCannotPay = errors.New("publisher's balance less than price * limit")
+	errCannotPay = errors.New("publisher's gas less than price * limit")
 )
 
 //go:generate mockgen -destination mock/engine_mock.go -package mock github.com/iost-official/go-iost/vm Engine
